% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_folder.R
\name{clone_folder}
\alias{clone_folder}
\title{Clone a set of github repos into a folder}
\usage{
clone_folder(selected, dest_dir, postfix = "submissions")
}
\arguments{
\item{selected}{list of repos as returned from `get_org_repos`}

\item{dest_dir}{Destination directory (default: working directory)}

\item{postfix}{user defined ending for folders, defaults to "-submissions"}
}
\description{
Clone a set of github repos into a folder
}
\examples{
\dontrun{
  ghc("451", clone_folder)
  }
}
