% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_github_repo.R
\name{clone_github_repo}
\alias{clone_github_repo}
\title{Clone a GitHub repository into the working directory}
\usage{
clone_github_repo(html, dest_dir = getwd())
}
\arguments{
\item{html}{Repository url}

\item{dest_dir}{Destination directory (default: working directory)}
}
\description{
Clone a GitHub repository into the working directory
}
