% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_repos.R
\name{distance_matrix}
\alias{distance_matrix}
\title{Pairwise distance matrix for strings}
\usage{
distance_matrix(x, method, ...)
}
\arguments{
\item{x}{character vector}

\item{method}{name of the method}

\item{...}{additional parameters passed on to the method}
}
\value{
numeric matrix of distances
}
\description{
Wrapper for the `stringdist::stringdist` function that calculates
distance between a pair of characters
}
\examples{
distance_matrix(c("statistics", "states", "status"), method="jw")
}
