\name{GHC}
\alias{GHC}
\docType{package}
\title{
This package includes several methods that can be used to cluster categorical data.
}
\description{
EnsCat implements several methods for the clustering of categorical data.}
\details{
\tabular{ll}{
Package: \tab EnsCat\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-22\cr
License: \tab GPL-2\cr
}
}
\author{
Maintainer: Saeid Amiri <saeid.amiri1@gmail.com>
}
\references{
Amiri, S., Clarke, B. & Clarke J. (2015). Clustering categorical data via ensembling dissimilarity matrices.  arXiv:1506.07930.

\url{https://sites.google.com/site/saeidamiri1/codes/clusofcate}

}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%#~~ simple examples of the most important functions ~~
%}
