% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_repos.R
\name{get_org_repos}
\alias{get_org_repos}
\title{Get all repositories for a GitHub organization}
\usage{
get_org_repos(org, visibility = "private", type = "all", per_page = 100)
}
\arguments{
\item{org}{Character string. GitHub organization name.}

\item{visibility}{One of "all", "public", or "private".}

\item{type}{One of "all", "sources", "forks", or "member".}

\item{per_page}{Number of results per page (max 100).}
}
\value{
A list of repository objects as returned by the GitHub API.
}
\description{
Get all repositories for a GitHub organization
}
\examples{
repos <- get_org_repos("stat451-at-unl")
length(repos)
}
