\name{GHC}
\alias{GHC}
\docType{package}
\title{
This package includes  methods that can be used to cluster of continuous data.
}
\description{
GHC implements methods for the clustering of continuous data.}
\details{
\tabular{ll}{
Package: \tab GHC\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-10-15\cr
License: \tab GPL-2\cr
}
}
\author{
Maintainer: Saeid Amiri <saeid.amiri1@gmail.com>
}
\references{
Amiri, S., Clarke, B, & Clarke, J. (2018). A General Hybrid Clustering Technique. To appear in Journal of Computational and Graphical Statistics.

\url{https://github.com/saeidamiri1/GHC}

}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%#~~ simple examples of the most important functions ~~
%}
