% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_pushed.R
\name{last_pushed}
\alias{last_pushed}
\title{Order a set of repos by most recent change}
\usage{
last_pushed(selected, dest_dir)
}
\arguments{
\item{selected}{list of repos as returned from `get_org_repos`}

\item{dest_dir}{Destination directory (default: working directory)}
}
\value{
data frame of repositories arranged by date and time of the last push
}
\description{
Order a set of repos by most recent change
}
