% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_repos.R
\name{longest_common_prefix}
\alias{longest_common_prefix}
\title{Longest common sprefix}
\usage{
longest_common_prefix(x)
}
\arguments{
\item{x}{Character vector (length >= 1)}
}
\value{
Character string giving the longest common prefix
}
\description{
Longest common sprefix
}
\examples{
longest_common_prefix(c("statistics", "state", "station"))
longest_common_prefix(c("apple", "banana"))
}
