% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org_repos.R
\name{longest_common_substring}
\alias{longest_common_substring}
\title{Longest common substring across multiple strings}
\usage{
longest_common_substring(x)
}
\arguments{
\item{x}{Character vector (length >= 2)}
}
\value{
Character string: longest common substring
}
\description{
Longest common substring across multiple strings
}
\examples{
longest_common_substring(c("statistics", "states", "taste"))
longest_common_substring(c("abcdef", "zabcf", "tabcxy"))
}
