% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_folder.R
\name{pull_folder}
\alias{pull_folder}
\title{Order a set of repos by most recent change}
\usage{
pull_folder(selected, dest_dir, postfix = "submissions")
}
\arguments{
\item{selected}{list of repos as returned from `get_org_repos`}

\item{dest_dir}{Destination directory (default: working directory)}

\item{post_fix}{user defined ending for folders, defaults to "-submissions"}
}
\value{
data frame of repositories arranged by date and time of the last push
}
\description{
Order a set of repos by most recent change
}
