% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_repos.R
\name{ghc}
\alias{ghc}
\title{Identify a set of repos from an organization}
\usage{
ghc(course = "451", f)
}
\arguments{
\item{course}{organization to check}

\item{f}{function to be called with the selected repos}
}
\value{
function call to f with the identified repos
}
\description{
Identify a set of repos from an organization
}
